requirements.txt
streamlit
langchain-google-genai
langgraph
langchain-core
gitpython
chromadb
sentence-transformers
python-dotenv